requirements.txt
# StructGAN Project Dependencies
# Deep Learning Frameworks
torch>=1.9.0
torchvision>=0.10.0

# Image Processing
opencv-python>=4.5.5
scikit-image>=0.18.3
Pillow>=8.4.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.3
seaborn>=0.11.2
tensorboard>=2.7.0

# Web Interface
streamlit>=1.12.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
dominate>=2.6.0
requests>=2.26.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0
